thoughts and suggestions
Some thought on the articles in this section, will report them here again:

I am going through the handbook, article by article, This specific one bothered me a bit. 
It seems that, by exclusion, you are telling founders to only take advice from 1) Costumers (which sounds exactly right)  or 2) You or Investors just like you, not other types of investors, not other types of investors who have built successful startups, but Investors in the Mena regions, who have built previously successful startups.  

If this is indeed the message that you are going for, why not boldly state it, and state the reason for it.

To a lesser extent, the article 'When to raise money', also heavily seems to give the idea, that founders should just make it easier for investors, and accept whatever scraps they are given. 

This seems a bit dishonest, to be frank, why would a Startup founder that can effectively bootstrap for years until they have a successful idea ever take investment? Why not take a loan? Or just keep going as they are until they become profitable? Or if they are looking to aggressively expand, just directly go into foreign investment? 
It's very worrying that when it comes to Investment, you ask of founders to just let you in, what exactly is Maza adding to a successful startup at that point?

Some harsh points and maybe a few criticism, but I am sure many founders have previously asked them as well, and maybe this article (or others) could be edited to further expand on them.

Especially the point of "Why not just bootstrap it"